Added a pause menu
Added a pause menu in the game using "Esc" key. You can resume by pressing it again or pressing the resume button. Pressing the quit button moves you back to the main menu.

The solution to block raycasts is pretty shit but i think this is the best there is, sorry.